Add files via upload
This is the version containing 15 colours, with viscosity and probability for each type of interaction and the strength and range of attractive and repulsive forces are between two user-defined limits (minimum and maximum). Randomization is not within the limits imposed by the code but within the limits given by the user.